LayoutTests/svg/dom/SVGTransformList-expected.txt

This test checks the SVGTransformList API

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Check passing invalid arguments to 'createSVGTransformFromMatrix'
PASS transform.createSVGTransformFromMatrix() threw exception SyntaxError: Not enough arguments.
PASS transform.createSVGTransformFromMatrix(svgElement.createSVGTransform()) threw exception TypeError: Type error.
PASS transform.createSVGTransformFromMatrix(svgElement) threw exception TypeError: Type error.
PASS transform.createSVGTransformFromMatrix('aString') threw exception TypeError: Type error.
PASS transform.createSVGTransformFromMatrix(1) threw exception TypeError: Type error.
PASS transform.createSVGTransformFromMatrix(true) threw exception TypeError: Type error.
PASS successfullyParsed is true

TEST COMPLETE
